+++ Work Keeper/Assets.xcassets/Custom Icons/inProgress.imageset/Contents.json
@@ -1,17 +1,50 @@
 {
   "images" : [
     {
-      "filename" : "inProgress.svg",
+      "filename" : "schedule.svg",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
-      "filename" : "inProgress 1.svg",
+      "appearances" : [
+        {
+          "appearance" : "luminosity",
+          "value" : "dark"
+        }
+      ],
+      "filename" : "schedule dark.svg",
+      "idiom" : "universal",
+      "scale" : "1x"
+    },
+    {
+      "filename" : "schedule 1.svg",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
-      "filename" : "inProgress 2.svg",
+      "appearances" : [
+        {
+          "appearance" : "luminosity",
+          "value" : "dark"
+        }
+      ],
+      "filename" : "schedule dark 1.svg",
+      "idiom" : "universal",
+      "scale" : "2x"
+    },
+    {
+      "filename" : "schedule 2.svg",
+      "idiom" : "universal",
+      "scale" : "3x"
+    },
+    {
+      "appearances" : [
+        {
+          "appearance" : "luminosity",
+          "value" : "dark"
+        }
+      ],
+      "filename" : "schedule dark 2.svg",
       "idiom" : "universal",
       "scale" : "3x"
     }
