--- Work Keeper/Modifications/Icons.swift
@@ -1,12 +1,0 @@
-import SwiftUI
-
-extension View {
-    @ViewBuilder
-    func ifAvailableButtonStyleGlass() -> some View {
-        if #available(iOS 26, *) {
-            self.buttonStyle(.glass)
-        } else {
-            self
-        }
-    }
-}
